requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Background Jobs
celery==5.3.6
croniter==2.0.1

# Caching
redis==5.0.1

# Email
fastapi-mail==1.4.1

# AI and Speech Processing
openai==1.3.5
google-cloud-speech==2.21.0
google-cloud-storage==2.13.0

# Monitoring and Metrics
prometheus-client==0.19.0

# Utils
pydantic==2.5.2
pydantic-settings==2.1.0
python-dateutil==2.8.2
tenacity==8.2.3
typer==0.9.0

# Testing (Optional but recommended)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools (Optional)
black==23.11.0
flake8==6.1.0
isort==5.12.0